+++ src/definitions/index.ts
@@ -1,3 +1,4 @@
+export * from './causa.js';
 export * from './emulator.js';
 export * from './environment.js';
 export * from './event-topic.js';